:bug: Cast durations

+++ src/clients/http.rs
@@ -392,7 +392,7 @@ impl OnResponse<Incoming> for ClientOnResponse {
             request_duration = latency.as_millis()
         );
         info!(
-            histogram.client_request_duration = latency.as_millis(),
+            histogram.client_request_duration = latency.as_millis() as u64,
             response_status = response.status().as_u16()
         );
 
@@ -490,7 +490,7 @@ impl OnEos for ClientOnEos {
             monotonic_counter.client_stream_response_count = 1,
             stream_duration = stream_duration.as_millis()
         );
-        info!(histogram.client_stream_response_duration = stream_duration.as_millis());
+        info!(histogram.client_stream_response_duration = stream_duration.as_millis() as u64);
     }
 }
